Working on a look ahead to chord tones tone to note converter.
Currently we generate all the chord tones first (choose a random chord
tone with probabilistically choosing the octave). Now we want to be
able to choose notes from the other tones based on the chord tones
around it. We should associate each other tone with a chord tone
(perhaps) and then we can use that chord tone to shape how we generate
the other tones.